Completed the funtina RemoveTriplesMap, now ask the user to remove the join conditions which the removed triples map is the parent triples map.
Changed the initialization of predicated object maps.

Now asks the user to provide the the triples map's rdf namespace.

Changed the way a r2rml map is created: now a database has to be loaded
in the sistem. When a new database is loaded a new r2rml map is created.